// src/app/interfaces/solicitudTransparencia.ts
export interface solicitudTransparencia {
  "NO": number;
  "UTAG": Date;
  "Folio": string;
  "R-Revisión": string;
  "tipo_Solicitud_id": number;
  "Solicitud": string;
  "estatus_solicitud_id": number;
  "Estatus": string;
  "Límite": Date;
  "Gracia": Date;
  "Transcurridos": number;
  "Solicitud_Terminada": Date;
  "tema_id": number;
  "subtema_id": number;
  "solicitud": string;
  "observaciones": string;
}