// src/app/interfaces/expotSolicitudes.ts
export interface exportSolicitudes {
    "NO": number;
    "UTAG": Date;
    "Folio": string;
    "R-Revisión": string;
    "Solicitud": string;
    "Estatus": string;
    "Fecha Límite": Date;
    "Fecha Gracia": Date;
    "Fecha de Termino": Date;
    "Días Hábiles Transcurridos": string;
    "solicitud": string;
    "observaciones": string;
}