// src/app/interfaces/exportSolicitudesDetalle.ts
export interface exportSolicitudesDetalle {
    "NO": number;
    "UTAG": Date;
    "Folio": string;
    "R-Revisión": string;
    "Solicitud": string;
    "Estatus": string;
    "Fecha Límite": Date;
    "Fecha Gracia": Date;
    "Fecha de Termino": Date;
    "Días Hábiles Transcurridos": string;
    "solicitud": string;
    "observaciones": string;
    "Tema": string;
    "Subtema": string;
    "Fecha": Date;
    "Asunto": string;
    "Remitente": string;
    "Correo": string;
    "Destinatario": string;
    "solicitud_dsc": string;
    "sol_Duplicada": number;
    "sol_Complementaria": number;
    "dias_habiles": number;
    "solicitudTer": string;
}